Added logic for generating pages for manage apprentices page
The main bulk of the happy path logic is in this commit. Controller level and below is covered with only the view logic left to add. Some error scenarios have yet to be covered as well and nuget package updates to support client updates are also needed for this code to work